FInished CLI Implementation

+++ mbroker/mbroker.c
@@ -14,7 +14,6 @@
 
 sem_t sessionsSem;
 pc_queue_t queue;
-Box_Node* box_list = NULL;
 
 int main(int argc, char** argv) {
   // expected argv:
@@ -30,18 +29,19 @@ int main(int argc, char** argv) {
   const char* pipeName = argv[1];
   const size_t maxSessions = (size_t)atoi(argv[2]);
 
+  pthread_t workerThreads[maxSessions];
+  for (int i = 0; i < maxSessions; i++) {
+    pthread_create(workerThreads[i], NULL, &workerThreadFunc, NULL);
+  }
+
   pcq_create(&queue, (size_t)maxSessions);
 
-  // if (sem_init(&sessionsSem, 0, maxSessions) == -1) {
-  //   perror("sem_init");
-  //   exit(EXIT_FAILURE);
-  // }
-
-  // sem_wait(&sessionsSem); //! Usar isto quando não houverem mais sessões
+  if (sem_init(&sessionsSem, 0, (unsigned int)maxSessions) == -1) {
+    perror("sem_init");
+    exit(EXIT_FAILURE);
+  }
 
   tfs_init(NULL);
-
-  box_list = NULL;
 
   unlink(pipeName);
 
@@ -59,58 +59,36 @@ int main(int argc, char** argv) {
   while (1) {
     Registry_Protocol* registry =
         (Registry_Protocol*)malloc(sizeof(Registry_Protocol));
-    // This loop reads the pipe, always expecting new messages
+    // This loop reads the pipe, always expecting new registrys
 
     if ((read(fd, registry, sizeof(Registry_Protocol))) != 0) {
-      fprintf(stdout, "ERROR %d\n", registry->code);
-      //* recebeu uma mensagem
+      // Received a registry
 
       switch (registry->code) {
         case 1:
-          register_pub(registry->register_pipe_name, registry->box_name,
-                       box_list);
+          register_pub(registry->register_pipe_name, registry->box_name);
           break;
 
         case 2:
-          register_sub(registry->register_pipe_name, registry->box_name,
-                       box_list);
+          register_sub(registry->register_pipe_name, registry->box_name);
           break;
 
         case 3:
-          create_box(registry->register_pipe_name, registry->box_name,
-                     box_list);
-          if (box_list == NULL) {
-            puts("Nula");
-          }
-          break;
-
-        case 4:
-          //* resposta
+          create_box(registry->register_pipe_name, registry->box_name);
           break;
 
         case 5:
-          destroy_box(registry->register_pipe_name, registry->box_name,
-                      box_list);
-          break;
-
-        case 6:
-          //* resposta
+          destroy_box(registry->register_pipe_name, registry->box_name);
           break;
 
         case 7:
-          puts("hey");
-          send_list_boxes(registry->register_pipe_name, box_list);
-          break;
-
-        case 8:
-          //* resposta
+          send_list_boxes(registry->register_pipe_name);
           break;
 
         default:
           break;
       }
     }
-    sleep(1);  //! Ta em espera ativa aqui uwu
     free(registry);
   }
 
@@ -119,6 +97,8 @@ int main(int argc, char** argv) {
   tfs_destroy();
 
   close(fd);
+
+  sem_destroy(&sessionsSem);
 
   return EXIT_FAILURE;
 }
